Extract out status endpoint
